+++ src/stormwind-library.lua
@@ -1,10 +1,19 @@
--- Library version = '0.0.6'
-
--- import src/Core/Factory.lua
-
--- import src/Facades/Target.lua
-
--- import src/Models/Macro.lua
+-- Library version = '0.0.7'
 
 -- import src/Support/Arr.lua
 -- import src/Support/Str.lua
+
+-- import src/Core/AddonProperties.lua
+-- import src/Core/Factory.lua
+-- import src/Core/Output.lua
+
+-- import src/Commands/Command.lua
+-- import src/Commands/CommandsHandler.lua
+
+-- import src/Facades/Events.lua
+-- import src/Facades/EventHandlers/PlayerLoginEventHandler.lua
+-- import src/Facades/EventHandlers/TargetEventHandler.lua
+-- import src/Facades/Target.lua
+
+-- import src/Models/Macro.lua
+-- import src/Models/RaidMarker.lua