+++ documentation/docs/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2024.mm.yy - version 0.0.7-alpha
+
+* Add the Events facade class to serve as a way to improve event handling
+    * Listen to `PLAYER_LOGIN`
+    * Listen to `PLAYER_TARGET_CHANGED` and split it into three events
+* Add the CommandsHandler class to allow addons to register commands
+* Add the Output class to replace print() calls
+* Add the RaidMarker model class and the library's possible marker instances
+* Allow passing addon properties to the library
+
 ## 2024.03.16 - version 0.0.6-alpha
 
 * Add the Str support class